opencti_rag_connector/src/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
arrow==1.3.0
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama==0.4.6
colour==0.1.5
cpe==1.3.1
cssselect==1.2.0
cybox==2.1.0.21
dataclasses-json==0.6.7
datefinder==0.7.3
deepdiff==8.0.1
Deprecated==1.2.15
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
drawsvg==2.4.0
elastic-transport==8.15.1
elasticsearch==8.16.0
et_xmlfile==2.0.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
filigran-sseclient==1.0.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
importlib-metadata==6.11.0
isoduration==20.11.0
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama-index-core==0.12.2
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.1
llama-index-readers-web==0.3.0
llama-index-vector-stores-elasticsearch==0.4.0
loguru==0.7.2
lxml==5.3.0
maec==4.1.0.17
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mitreattack-python==3.0.8
mixbox==1.0.5
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.56.1
openpyxl==3.1.5
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
orderly-set==5.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pika==1.3.2
pillow==11.0.0
platformdirs==4.3.6
playwright==1.49.0
pluralizer==1.2.0
pooch==1.8.2
prometheus_client==0.20.0
propcache==0.2.1
pycountry==24.6.1
pycti==6.4.2
pydantic==2.8.2
pydantic_core==2.20.1
pyee==12.0.0
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
sentence-transformers==3.3.1
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.0.27
SQLAlchemy==2.0.36
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.7
stix2-patterns==2.0.0
stix2-validator==3.2.0
stixmarx==1.0.8
sympy==1.13.1
tabulate==0.9.0
taxii2-client==2.3.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.15.0
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
weakrefmethod==1.0.3
webcolors==24.11.1
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.18.3
zipp==3.21.0
